requirements.txt
allure-behave==2.8.18
allure-pytest==2.8.29
allure-python-commons==2.8.29
apipkg==1.5
appdirs==1.4.4
Appium-Python-Client==1.0.2
astor==0.8.1
atomicwrites==1.4.0
attrs==20.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
behave==1.2.6
bs4==0.0.1
Calculator==1.1.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
configparser==5.0.1
crayons==0.4.0
cryptography==3.1.1
cucumber-tag-expressions==3.0.0
et-xmlfile==1.0.1
execnet==1.7.1
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
jdcal==1.4.1
lxml==4.6.1
more-itertools==8.5.0
mysql-connector-python==8.0.21
oops==0.0.14
openpyxl==3.0.5
packaging==20.4
paramiko==2.7.2
parse==1.18.0
parse-type==0.5.2
playsound==1.2.2
pluggy==0.13.1
protobuf==3.13.0
py==1.9.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
PySelenium==1.0.2
pytest==6.0.2
pytest-ast-transformer==1.0.3
pytest-forked==1.3.0
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-ordering==0.6
pytest-soft-assertions==0.1.2
pytest-xdist==2.1.0
pythonlogger==0.0.1
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
texttable==1.6.3
toml==0.10.1
tqdm==4.51.0
urllib3==1.25.10
wait==0.0.3
webdriver-manager==3.2.2
webdrivermanager==0.9.0
zipp==3.1.0
